주소 등록 구현중

+++ src/main/java/ga/todayOutside/utils/JwtService.java
@@ -21,7 +21,7 @@ public class JwtService {
      * @param userId
      * @return String
      */
-    public String createJwt(int userId) {
+    public String createJwt(Long userId) {
         Date now = new Date();
         return Jwts.builder()
                 .claim("userId", userId)
@@ -44,7 +44,7 @@ public class JwtService {
      * @return int
      * @throws BaseException
      */
-    public int getUserId() throws BaseException {
+    public Long getUserId() throws BaseException {
         // 1. JWT 추출
         String accessToken = getJwt();
         if (accessToken == null || accessToken.length() == 0) {
@@ -62,6 +62,6 @@ public class JwtService {
         }
 
         // 3. userId 추출
-        return claims.getBody().get("userId", Integer.class);
+        return claims.getBody().get("userId", Long.class);
     }
 }
